requirements.txt
openpyxl
gspread
arabic_reshaper
python-bidi
